backend/pyproject.toml
[project]
name = "prism-genomics"
version = "0.2.0"
description = "Polygenic Risk Intelligence for Secure Medicine — Pathogenic/Benign variant classification from DNA data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.129.0",
    "uvicorn>=0.40.0",
    "python-multipart>=0.0.22",
    "torch>=2.2.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.4.0",
    "cryptography>=46.0.5",
    "blake3>=1.0.8",
]

[project.scripts]
train = "scripts.train:main"
